requirement.txt
absl-py==0.8.0
albumentations==0.3.3
astor==0.8.0
astroid==2.4.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
catalyst==19.9.4
certifi==2020.4.5.1
chardet==3.0.4
crc32c==1.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
efficientnet-pytorch==0.6.3
entrypoints==0.3
future==0.18.2
gast==0.3.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
idna==2.8
imageio==2.5.0
imgaug==0.2.6
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
jsonpatch==1.24
jsonpointer==2.0
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
kaggle==1.5.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lz4==2.2.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
munch==2.3.2
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.1
numpy==1.17.2
opencv-python==4.1.1.26
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plotly==4.1.1
pretrainedmodels==0.7.4
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
pyarrow==0.14.1
Pygments==2.4.2
pylint==2.5.2
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
python-slugify==3.0.3
pytorch-toolbelt==0.1.4
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
retrying==1.3.3
safitty==1.2.5
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
segmentation-models-pytorch==0.0.1
Send2Trash==1.5.0
six==1.14.0
tensorboard==1.14.0
tensorboardX==1.8
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
toml==0.10.0
torch==1.2.0
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.2.2
tornado==6.0.3
tqdm==4.29.1
traitlets==4.3.2
ttach==0.0.2
typed-ast==1.4.1
urllib3==1.24.3
visdom==0.1.8.9
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.6
widgetsnbextension==3.5.1
wrapt==1.11.2
